+++ include/sw/dsp/sdr/sdr.hpp
@@ -10,3 +10,4 @@
 // SPDX-License-Identifier: MIT
 
 #include <sw/dsp/sdr/constellation.hpp>
+#include <sw/dsp/sdr/rrc.hpp>